Split project into dedicated files
