+++ ComposableArchExample/ComposableArchExampleApp.swift
@@ -7,9 +7,13 @@ struct ComposableArchExampleApp: App {
         WindowGroup {
             NavigationView {
                 NavigationLink("Navigation"){
-                    CounterView(store: Store(initialState: CounterViewState(),
-                                             reducer: reducer.debug(),
-                                             environment: CounterViewEnvironment(dispatchQueue: DispatchQueue.main)))
+                    CounterView(
+                        store: Store(
+                            initialState: CounterViewState(),
+                            reducer: reducer.debug(),
+                            environment: .init(queue: DispatchQueue.main)
+                        )
+                    )
                     
                 }.navigationTitle("Main view")
             }
