+++ ComposableArchExample/Counter/CounterView.swift
@@ -1,0 +1,40 @@
+import SwiftUI
+import ComposableArchitecture
+
+struct CounterView: View {
+    let store: Store<CounterViewState, CounterViewAction>
+    
+    var body: some View {
+        WithViewStore(self.store, observe: { $0 }) { viewStore in
+            VStack {
+                Text("Timer[\(viewStore.isTimerOn ? "ON" : "OFF" )]: \(viewStore.timer)")
+                    .padding()
+                Text("\(viewStore.counter) is \(viewStore.isEvenNumber ? "even" : "odd")")
+
+                HStack{
+                    Button("-") {
+                        viewStore.send(.decrementTapped)
+                    }
+                    Text("\(viewStore.counter)")
+                    Button("+") {
+                        viewStore.send(.incrementTapped)
+                    }
+                }
+            }
+            .onDisappear{ viewStore.send(.onDisappear) }
+        }
+    }
+}
+
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        CounterView(
+            store: Store(
+                initialState: CounterViewState(),
+                reducer: reducer,
+                environment: .init(queue: DispatchQueue.main)
+            )
+        )
+    }
+}
